+++ src/tasks/log-rotator.cc
@@ -31,6 +31,8 @@
 #include "core/registry.h"
 #include "core/residue-exception.h"
 #include "crypto/zlib.h"
+#include "extensions/pre-archive-extension.h"
+#include "extensions/post-archive-extension.h"
 #include "logging/log.h"
 #include "utils/utils.h"
 
@@ -152,7 +154,7 @@ LogRotator::RotateTarget LogRotator::createRotateTarget(const std::string& logge
 
         std::unordered_set<std::string> fileByLevel;
 
-        std::map<std::string, std::set<std::string>> levelsInFilename;
+        std::unordered_map<std::string, std::set<std::string>> levelsInFilename;
 
         std::lock_guard<std::recursive_mutex> l(logger->lock());
 
@@ -246,7 +248,7 @@ void LogRotator::rotate(const std::string& loggerId)
 
     const RotateTarget rotateTarget = createRotateTarget(loggerId);
 
-    std::map<std::string, std::string> files;
+    std::unordered_map<std::string, std::string> files;
     el::Logger* logger = el::Loggers::getLogger(loggerId, false);
 
     if (logger != nullptr) {
@@ -318,7 +320,7 @@ void LogRotator::rotate(const std::string& loggerId)
     m_archiveItems.push_back({loggerId, rotateTarget.destinationDir + el::base::consts::kFilePathSeperator + rotateTarget.archiveFilename, files});
 }
 
-void LogRotator::archiveAndCompress(const std::string& loggerId, const std::string& archiveFilename, const std::map<std::string, std::string>& files) {
+void LogRotator::archiveAndCompress(const std::string& loggerId, const std::string& archiveFilename, const std::unordered_map<std::string, std::string>& files) {
     if (files.empty()) {
         RLOG(INFO) << "No file to archive for [" << loggerId << "]";
         return;
@@ -326,6 +328,25 @@ void LogRotator::archiveAndCompress(const std::string& loggerId, const std::stri
 
     RLOG(INFO) << "Archiving for [" << loggerId << "] => [" << archiveFilename
                << "] containing " << files.size() << " file(s)";
+
+
+    if (!m_registry->configuration()->preArchiveExtensions().empty()) {
+        PreArchiveExtension::Data d {
+            loggerId,
+            archiveFilename,
+            files
+        };
+        bool continueProcess = true;
+        for (auto& ext : m_registry->configuration()->preArchiveExtensions()) {
+            auto extResult = ext->trigger(&d);
+            continueProcess = continueProcess && extResult.continueProcess;
+        }
+
+        if (!continueProcess) {
+            RLOG(INFO) << "Stopping archive process for logger [" << loggerId << "] because of one of the [pre_archive] extensions";
+            return;
+        }
+    }
 
     // compress files after logger's lock is released
     RVLOG(RV_DETAILS) << "Compressing rotated files for logger [" << loggerId << "] to [" << archiveFilename << "]";
@@ -355,6 +376,17 @@ void LogRotator::archiveAndCompress(const std::string& loggerId, const std::stri
         if (::remove(tmpTar.c_str()) != 0) {
             RLOG(WARNING) << "Compressed file successfully but failed to remove tar: " << std::strerror(errno);
         }
+
+        if (!m_registry->configuration()->postArchiveExtensions().empty()) {
+            PostArchiveExtension::Data d {
+                loggerId,
+                archiveFilename
+            };
+            for (auto& ext : m_registry->configuration()->postArchiveExtensions()) {
+                ext->trigger(&d);
+            }
+        }
+
     } else {
         RLOG(ERROR) << "Failed to compress rotated log for logger [" << loggerId << "]. Destination name: [" << archiveFilename + "]";
     }
@@ -417,7 +449,7 @@ types::Time DailyLogRotator::calculateRoundOff(types::Time now) const
 types::Time WeeklyLogRotator::calculateRoundOff(types::Time now) const
 {
 
-    const std::map<std::string, int> WEEK_DAYS_MAP = {
+    const std::unordered_map<std::string, int> WEEK_DAYS_MAP = {
         { "Mon", 1 },
         { "Tue", 2 },
         { "Wed", 3 },
@@ -446,7 +478,7 @@ types::Time WeeklyLogRotator::calculateRoundOff(types::Time now) const
 types::Time MonthlyLogRotator::calculateRoundOff(types::Time now) const
 {
     const int year = atoi(Utils::formatTime(now, "%Y").c_str());
-    const std::map<int, int> LAST_DAY_OF_MONTH_MAP = {
+    const std::unordered_map<int, int> LAST_DAY_OF_MONTH_MAP = {
         { 1, 31 },
         { 2, year % 4 == 0 ? 29 : 28 },
         { 3, 31 },
@@ -485,7 +517,7 @@ types::Time MonthlyLogRotator::calculateRoundOff(types::Time now) const
 types::Time YearlyLogRotator::calculateRoundOff(types::Time now) const
 {
     const int year = atoi(Utils::formatTime(now, "%Y").c_str());
-    const std::map<int, int> LAST_DAY_OF_MONTH_MAP = {
+    const std::unordered_map<int, int> LAST_DAY_OF_MONTH_MAP = {
         { 1, 31 },
         { 2, year % 4 == 0 ? 29 : 28 },
         { 3, 31 },